+++ Versions/Temporary/Engine/Sources/ED_B2_M1/MapInfoAIMarkersSelectDlg.cpp
@@ -80,7 +80,7 @@ void CMapInfoAIMarkersSelectDlg::OnCancel()
 	CDialog::OnCancel();
 }
 
-bool CMapInfoAIMarkersSelectDlg::HandleCommand( unsigned nCommandID, uint32_t dwData )
+bool CMapInfoAIMarkersSelectDlg::HandleCommand( unsigned nCommandID, uintptr_t dwData )
 {
 	return false;
 }
