+++ Versions/Temporary/Engine/Sources/ED_B2_M1/MapObjectWindow.cpp
@@ -370,7 +370,7 @@ void CMapObjectWindow::UpdateSelection()
 							//
 							ICommandHandlerContainer* pCommandHandlerContainer = Singleton<ICommandHandlerContainer>();
 							pCommandHandlerContainer->Set( CHID_OBJECT_STORAGE, this );
-							pCommandHandlerContainer->HandleCommand( CHID_MAPINFO_MAPOBJECT_MULTI_STATE, ID_MIMO_SWITCH_MULTI_STATE, reinterpret_cast<uint32_t>( &( selectedObjectListElement.szObjectTypeName ) ) );
+							pCommandHandlerContainer->HandleCommand( CHID_MAPINFO_MAPOBJECT_MULTI_STATE, ID_MIMO_SWITCH_MULTI_STATE, reinterpret_cast<uintptr_t>( &( selectedObjectListElement.szObjectTypeName ) ) );
 							pCommandHandlerContainer->HandleCommand( CHID_MAPINFO_MAPOBJECT_STATE, ID_MIMO_SWITCH_ADD_STATE, 0 );	
 						}
 					}
@@ -396,7 +396,7 @@ void CMapObjectWindow::ClearSelection()
 }
 
 
-bool CMapObjectWindow::HandleCommand( unsigned nCommandID, uint32_t dwData )
+bool CMapObjectWindow::HandleCommand( unsigned nCommandID, uintptr_t dwData )
 {
 	switch( nCommandID )
 	{
@@ -455,7 +455,7 @@ bool CMapObjectWindow::HandleCommand( unsigned nCommandID, uint32_t dwData )
 				pObjectManipulator = pMultiManipulator;
 			}
 			IView *pView = 0;
-			Singleton<ICommandHandlerContainer>()->HandleCommand( CHID_PC_DIALOG, ID_PC_DIALOG_GET_VIEW, reinterpret_cast<uint32_t>( &pView ) );
+			Singleton<ICommandHandlerContainer>()->HandleCommand( CHID_PC_DIALOG, ID_PC_DIALOG_GET_VIEW, reinterpret_cast<uintptr_t>( &pView ) );
 			if ( pView != 0 )
 			{
 				pView->SetViewManipulator( pObjectManipulator, objectSet, std::string() );
