+++ Versions/Temporary/Engine/Sources/ED_B2_M1/ModelEditor.cpp
@@ -179,7 +179,7 @@ void CModelEditor::Destroy()
 }
 
 
-bool CModelEditor::HandleCommand( unsigned nCommandID, uint32_t dwData )
+bool CModelEditor::HandleCommand( unsigned nCommandID, uintptr_t dwData )
 {
 	switch( nCommandID ) 
 	{
