+++ Versions/Temporary/Engine/Sources/ED_B2_M1/RiverState.cpp
@@ -29,7 +29,7 @@ bool CRiverState::CanInsertVSO()
 	if ( CanEdit() )
 	{
 		SObjectSet objectSet;
-		if ( Singleton<ICommandHandlerContainer>()->HandleCommand( CHID_OBJECT_STORAGE, ID_OS_GET_OBJECTSET, reinterpret_cast<uint32_t>( &objectSet ) ) && ( !objectSet.objectNameSet.empty() ) )
+		if ( Singleton<ICommandHandlerContainer>()->HandleCommand( CHID_OBJECT_STORAGE, ID_OS_GET_OBJECTSET, reinterpret_cast<uintptr_t>( &objectSet ) ) && ( !objectSet.objectNameSet.empty() ) )
 		{
 			if( objectSet.szObjectTypeName == GetVSOTypeName() )
 			{
@@ -75,7 +75,7 @@ void CRiverState::UpdateVisualVSO( NDb::SVSOInstance *pVSO, bool bBothEdges )
 		{
 			float fDefaultWidth = 0.0f;
 			SObjectSet objectSet;
-			if ( Singleton<ICommandHandlerContainer>()->HandleCommand( CHID_OBJECT_STORAGE, ID_OS_GET_OBJECTSET, reinterpret_cast<uint32_t>( &objectSet ) ) && ( !objectSet.objectNameSet.empty() ) )
+			if ( Singleton<ICommandHandlerContainer>()->HandleCommand( CHID_OBJECT_STORAGE, ID_OS_GET_OBJECTSET, reinterpret_cast<uintptr_t>( &objectSet ) ) && ( !objectSet.objectNameSet.empty() ) )
 			{
 				if ( objectSet.szObjectTypeName == VSO_TYPE_NAME )
 				{
